Moved cli files to src folder since root is cluttered with configuration files. Refactored queryHelper into a exported util function. No need for it to be a class. Added texting to it
