+++ package.json
@@ -2,16 +2,16 @@
   "name": "imdb-scraper",
   "version": "2.1.1",
   "description": "Search and scrape IMDb for IMDb ID's",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "jest",
     "test:watch": "jest --watch",
-    "start": "node index.js",
+    "start": "node src/index.js",
     "lint": "eslint --ext .js --ignore-path .gitignore .",
     "lint:fix": "eslint --fix --ext .js --ignore-path .gitignore ."
   },
   "bin": {
-    "imdb": "./index.js"
+    "imdb": "./src/index.js"
   },
   "author": "Daniel Vernberg",
   "license": "ISC",
